New static and factory creators.
Added `AbstractLoomFactory` and `LoomFactory` as creators for the Loom object.
Moved `Loom->diff()` into the `LoomComparisons` trait.
Added a `daysPerMonth` parameter to `Loom->getMonths()` and to `LoomFactory->fromMonths()`
The `Months->__construct()` will also accept a `daysPerMonth` parameter.
Added a static `make()` method to the `Loom` object which returns a new instance of `LoomFactory`.